app/requirements.txt
mypy
black
isort
sqlalchemy
alembic
openai
python-dotenv
heyoo
fastapi
uvicorn
stripe
googlemaps
bcrypt
pyjwt
